Fix glitchy frames bug; Add setting for exposure time
Previously, there was a bug where Bonsai was being told that a
frame was fully copied and ready to send to the next node when
it wasn't. This led to Bonsai sending along frames that weren't
complete. Now, Bonsai only sends a frame to the next node when the
frame is fully copied from the camera.

The exposure time setting currently needs to be set before running
a workflow. This needs to be updated to allow real-time adjustment.